requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==3.1.7
djangorestframework==3.15.1
dotenv==0.9.9
joblib==1.3.2
numpy==1.21.6
pandas==1.3.5
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
pytz==2025.2
scikit-learn==1.0.2
scipy==1.7.3
six==1.17.0
sqlparse==0.4.4
threadpoolctl==3.1.0
typing_extensions==4.7.1
